+++ release_notes.md
@@ -1,11 +1,30 @@
 <!-- DO NOT MODIFY this file, it is generated automatically from dev_setup ! -->
 # SPIKE Relase Notes
+#### 0.99.23 - May 2021
+- first interactive notebook for phasing of FTICR-MS spectra - still a bit rough
+- slight improvements in apmin() (NMR automatic phaser) better search algorithm
+- slight improvements in bcorr() (baseline correction), when working on complex spectra
+- new doc is still in development and missing, sorry (phasing is keeping me busy) 
+
+#### 0.99.22 - April 2021 - not released
+- change in the plugin set up, there can now be distributed in several directories (in that order)
+    - `(distrib dir)/spike/plugins`  - basic plugins allways loaded
+    - `$HOME/spike/plugins`  - plugins specific to the user allways loaded
+    - `(distrib dir)/spike/plugins/NMR`  - plugins specific to NMR, loaded with `import spike.NMR`
+    - `(distrib dir)/spike/plugins/MS`  - plugins specific to MS, loaded with `import spike.FTMS`
+    - and as before, plugins with a name starting with a `_` are not loaded
+- new `PhaseMS` plugin, which implements quadratic phase correction and permits to phase FTICR spectra
+- a .tm() apodisation (trapeze) which emulates D.Kilgour apodisation for phase sensitive FTICR-MS
+- rewrote and reorganized  README and documentation
+
 #### 0.99.21 - Feb 2021
 - plugin which implements .diagonal() for computing the diagonal of 2D FTICR spectra 
+- changed the logic to generate downsampled 2D FTICR spectra - smaller files, smaller vignettes
 - small correction in urQRd  - thanks to Will Kew
 - correction when reading Apex MS dataset for pulse frequency limits - thanks to Maria van Agthoven
 - correction ThermoFisher/Orbitrap import code - thanks to Will Kew
 - small corrections when opening files
+- many small bugs corrected
 
 #### 0.99.20 - Nov 2020
 - corrected a bad bug which corrupted F1 calibration when loading a 2D-FTICRMS experiment